[fix]Folder restructure into pages and components
