+++ src/App.js
@@ -3,23 +3,21 @@ import { Layout } from "antd";
 import "./antd.css";
 import "./design system/primary-styles.css";
 import "./App.css";
+import { FooterComponent, Navbar, NewPostModal } from "./components";
+import { ToastContainer } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css";
+import { RequiresAuth } from "./utils";
 import {
   BookmarkPage,
   ErrorPage,
   ExplorePage,
-  Feed,
-  FooterComponent,
+  FeedPage,
   LoginPage,
-  Navbar,
-  NewPostModal,
-  ProfilePage,
-  UserPage,
-  NotificationPage,
   OtherUserPage,
-} from "./components";
-import { ToastContainer } from "react-toastify";
-import "react-toastify/dist/ReactToastify.css";
-import { RequiresAuth } from "./utils";
+  ProfilePage,
+  NotificationPage,
+  UserPage,
+} from "./pages";
 
 function App() {
   return (
@@ -40,7 +38,7 @@ function App() {
           >
             <Route path="/user/explore" element={<ExplorePage />} />
             <Route path="/user/bookmark" element={<BookmarkPage />} />
-            <Route path="/user/feed" element={<Feed />} />
+            <Route path="/user/feed" element={<FeedPage />} />
             <Route path="/user/profile" element={<ProfilePage />} />
             <Route path="/user/notifications" element={<NotificationPage />} />
             <Route path="/user/:userID" element={<OtherUserPage />} />
