+++ src/components/modals/delete-post-modal.jsx
@@ -4,6 +4,7 @@ import "./modals.css";
 import { deletePost, getPosts } from "../../firebase/firestore-methods";
 import { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
+import { toast } from "react-toastify";
 
 const DeletePostModal = ({ isVisible, toggleModal, postID }) => {
   const { token } = useSelector((store) => store.token);
@@ -16,7 +17,10 @@ const DeletePostModal = ({ isVisible, toggleModal, postID }) => {
       await deletePost(postID, token);
       dispatch(getPosts());
       toggleModal();
-    } catch (err) {}
+      toast.success("Post Deleted");
+    } catch (err) {
+      toast.error("Post Delete Failed")
+    }
     setIsLoading(false);
   };
