+++ src/components/navbar/search/search-result.jsx
@@ -1,0 +1,3 @@
+const SearchResult = () => {};
+
+export { SearchResult };
