+++ src/pages/index.jsx
@@ -1,0 +1,9 @@
+export { FeedPage } from "./feed-page/feed-page";
+export { LoginPage } from "./auth-pages/login-page";
+export { BookmarkPage } from "./bookmark-page/bookmark";
+export { ErrorPage } from "./error-page/error-page";
+export { ExplorePage } from "./explore-page/explore";
+export { NotificationPage } from "./notification-page/notification-page";
+export { ProfilePage } from "./profile-page/profile-page";
+export { OtherUserPage } from "./profile-page/other-user-page";
+export { UserPage } from "./user-page/user-page";
